AND OPERATOR : ( true && false )
It will print false.

If on left side it is true the interpreter will check on the left and print whatever is present on right on screen.

If on left side it is other than true/valid string/number it will print as it is which is written on left side.